+++ satella/coding/predicates/__init__.py
@@ -1,0 +1,4 @@
+from .number import between
+from .generic import one_of
+
+__all__ = ['between', 'one_of']