+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.9.24
 
+* renamed an argument in `walk`
+