+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.9.24
 
 * renamed an argument in `walk`
-
+* added `update_key_if_not_none`
+* deprecated `update_key_if_none`